msp-cli/pyproject.toml
[tool.poetry]
name = "msp-cli"
version = "0.1.0"
description = "MSP CLI для управления N8N workflow"
authors = ["Ваше Имя <email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "0.9.0"
click = "8.1.7"
pydantic = "^2.0"
httpx = "^0.27"
jinja2 = "^3.1"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
ruff = "^0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"